Update base api class name
